Support multiple service ports
But still only a single port from consul

This is an attempt to work around ports missing when applying a service
which changes the port names. If the name changes then cube doesn’t know
where to send traffic. If 2+ ports can be used in the service then there
is a chance to have a working port while migrating. Still not great, but
better than no option at all