kube: Only attempt to sync IPv4 addresses
Kube will fail if a non-IP is used when attempting to sync the endpoint.

This otherwise causes a problem when an ELB is registered into Consul and
that service is sync’d back into Kube.

It would be possible to resolve the addresses and add those to Kube. This
may be done in the future or available as an option.